+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 # Version 2.3.0
 
+* Session view now displays durations
 * Sessions and tests now have a preview of the latest comment made on them, when you hover over them. You can also hit `c` to toggle showing the latest comment on all items in view
 
 # Version 2.2.0